requirements.txt
appdirs==1.4.2
Django==1.10.6
django-allauth==0.31.0
django-braces==1.11.0
django-invitations==1.9
oauthlib==2.0.2
packaging==16.8
pyparsing==2.2.0
python-openid==2.2.5
requests==2.13.0
requests-oauthlib==0.8.0
six==1.10.0
